+++ ElectronicObserver.KancolleApi.Types/ApiReqSortie/Models/APIStage3.cs
@@ -15,7 +15,7 @@ public class ApiStage3
 	[JsonPropertyName("api_edam")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<int> ApiEdam { get; set; } = new();
+	public List<double> ApiEdam { get; set; } = new();
 
 	[JsonPropertyName("api_erai_flag")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
@@ -35,7 +35,7 @@ public class ApiStage3
 	[JsonPropertyName("api_fdam")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<int> ApiFdam { get; set; } = new();
+	public List<double> ApiFdam { get; set; } = new();
 
 	[JsonPropertyName("api_frai_flag")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
