+++ ElectronicObserver.KancolleApi.Types/Models/APIOpeningAtack.cs
@@ -10,7 +10,7 @@ public class ApiOpeningAtack
 	[JsonPropertyName("api_edam")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<int> ApiEdam { get; set; } = new();
+	public List<double> ApiEdam { get; set; } = new();
 
 	[JsonPropertyName("api_erai")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
@@ -30,7 +30,7 @@ public class ApiOpeningAtack
 	[JsonPropertyName("api_fdam")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<int> ApiFdam { get; set; } = new();
+	public List<double> ApiFdam { get; set; } = new();
 
 	[JsonPropertyName("api_frai")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
